pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "discobase"
readme = "README.md"
license = "MIT"
dependencies = ["discord.py"]
dynamic = ["version"]

[tool.ruff.lint]
ignore = ["F403"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
